parking-notify/backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
qrcode[pil]==7.3.1
cryptography==36.0.2
python-dotenv==0.19.2
requests==2.27.1
